put request helper methods in http.py
